Little fixes
I was going through my old projects cause I can't leave well enough alone, and realized this class had no description and no author. If code is written and no one is around to claim it, does it make a sound?
Figured it was better to fix it than leave it.
Also added if statement to draw method to make boolean variable relevant.